type of va_arg promoted to 'int'

+++ firmware/u2f-stm32/Src/bsp.c
@@ -121,7 +121,7 @@ void u2f_printd(const char * tag, uint8_t c, ...)
     va_start(args,c);
     while(c--)
     {
-        u2f_putd((int32_t)va_arg(args, int16_t));
+        u2f_putd((int32_t)va_arg(args, int));
 
     }
     put_line();
@@ -149,7 +149,7 @@ void u2f_printx(const char * tag, uint8_t c, ...)
     va_start(args,c);
     while(c--)
     {
-        u2f_putx((int32_t)va_arg(args, uint16_t));
+        u2f_putx((int32_t)va_arg(args, int));
         u2f_prints(" ");
     }
     put_line();
@@ -163,7 +163,7 @@ void u2f_printb(const char * tag, uint8_t c, ...)
     va_start(args,c);
     while(c--)
     {
-        u2f_putb(va_arg(args, uint8_t));
+        u2f_putb(va_arg(args, int));
         put_space();
     }
     put_line();
